+++ odbc-api/src/handles/sql_result.rs
@@ -75,6 +75,7 @@ impl ExtSqlReturn for SqlReturn {
             SqlReturn::SUCCESS_WITH_INFO => SqlResult::SuccessWithInfo(()),
             SqlReturn::ERROR => SqlResult::Error { function },
             SqlReturn::NO_DATA => SqlResult::NoData,
+            SqlReturn::NEED_DATA => SqlResult::NeedData,
             SqlReturn::STILL_EXECUTING => SqlResult::StillExecuting,
             r => panic!(
                 "Unexpected return value '{:?}' for ODBC function '{}'",