+++ Changelog.md
@@ -10,6 +10,8 @@
 * `handles::Environment::drivers_buffer_fill` now returns `SqlResult<()>` instead of `SqlResult<bool>`.
 * `handles::Environment::drivers_buffer_len` now returns `SqlResult<(i16, i16)>` instead of `SqlResult<Option<(i16, i16)>>`.
 * `handles::Environment::data_source_buffer_len` now returns `SqlResult<(i16, i16)>` instead of `SqlResult<Option<(i16, i16)>>`.
+* `handles::Connection::execute` now returns now returns `SqlResult<()>` instead of `SqlResult<bool>`.
+* `handles::Connection::exec_direct` now returns now returns `SqlResult<()>` instead of `SqlResult<bool>`.
 * `handles::Statement::fetch` now returns now returns `SqlResult<()>` instead of `SqlResult<bool>`.
 
 ### Non Breaking